refactor: 문장 게임 패키지 구조 수정
