+++ src/main/java/com/capstone/storyforest/sentencegame/creativity/CreativityService.java
@@ -1,0 +1,5 @@
+package com.capstone.storyforest.sentencegame.creativity;
+
+public interface CreativityService {
+    boolean isCreative(String sentence);
+}
