+++ src/main/java/com/capstone/storyforest/sentencegame/service/OpenAiProfanityService.java
@@ -1,7 +1,7 @@
 package com.capstone.storyforest.sentencegame.service;
 
-import com.capstone.storyforest.filtering.dto.CurseCheckResult;
-import com.capstone.storyforest.filtering.service.AiCheckManager;
+import com.capstone.storyforest.sentencegame.filtering.dto.CurseCheckResult;
+import com.capstone.storyforest.sentencegame.filtering.service.AiCheckManager;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
